Update block-reorder.js
I have a Block containing a YouTube Video. When one tries to move it, the dragging stops because of some problems with the JS and the iFrame of the Video. It would be nice if the blocks can prepare themselves to be moved, or even als for a confirmation of the order of the blocks is of any meaning to the content - story telling.